+++ test/session.test.js
@@ -660,13 +660,13 @@ describe(`session: ${printPath("[test/session.test.js]")}`, function () {
         let res = await SessionFunctions.createNewSession(s.helpers, "", false, {}, {});
         await SessionFunctions.updateSessionDataInDatabase(s.helpers, res.session.handle, { key: "value" });
 
-        let res2 = (await SessionFunctions.getSessionInformation(s.helpers, res.session.handle)).sessionData;
+        let res2 = (await SessionFunctions.getSessionInformation(s.helpers, res.session.handle)).sessionDataInDatabase;
         assert.deepStrictEqual(res2, { key: "value" });
 
         //changing the value of session data with the same key
         await SessionFunctions.updateSessionDataInDatabase(s.helpers, res.session.handle, { key: "value 2" });
 
-        let res3 = (await SessionFunctions.getSessionInformation(s.helpers, res.session.handle)).sessionData;
+        let res3 = (await SessionFunctions.getSessionInformation(s.helpers, res.session.handle)).sessionDataInDatabase;
         assert.deepStrictEqual(res3, { key: "value 2" });
 
         //passing invalid session handle when updating session data
@@ -701,13 +701,13 @@ describe(`session: ${printPath("[test/session.test.js]")}`, function () {
         await SessionFunctions.updateSessionDataInDatabase(s.helpers, res.session.handle, { key: "value" });
 
         let res2 = await SessionFunctions.getSessionInformation(s.helpers, res.session.handle);
-        assert.deepStrictEqual(res2.sessionData, { key: "value" });
+        assert.deepStrictEqual(res2.sessionDataInDatabase, { key: "value" });
 
         //changing the value of session data with the same key
         await SessionFunctions.updateSessionDataInDatabase(s.helpers, res.session.handle, { key: "value 2" });
 
         let res3 = await SessionFunctions.getSessionInformation(s.helpers, res.session.handle);
-        assert.deepStrictEqual(res3.sessionData, { key: "value 2" });
+        assert.deepStrictEqual(res3.sessionDataInDatabase, { key: "value 2" });
 
         //passing invalid session handle when updating session data
         assert(!(await SessionFunctions.updateSessionDataInDatabase(s.helpers, "random", { key2: "value2" })));
@@ -731,27 +731,27 @@ describe(`session: ${printPath("[test/session.test.js]")}`, function () {
         //adding session data
         let res = await SessionFunctions.createNewSession(s.helpers, "", false, {}, null);
 
-        let res2 = (await SessionFunctions.getSessionInformation(s.helpers, res.session.handle)).sessionData;
+        let res2 = (await SessionFunctions.getSessionInformation(s.helpers, res.session.handle)).sessionDataInDatabase;
         assert.deepStrictEqual(res2, {});
 
         await SessionFunctions.updateSessionDataInDatabase(s.helpers, res.session.handle, { key: "value" });
 
-        let res3 = (await SessionFunctions.getSessionInformation(s.helpers, res.session.handle)).sessionData;
+        let res3 = (await SessionFunctions.getSessionInformation(s.helpers, res.session.handle)).sessionDataInDatabase;
         assert.deepStrictEqual(res3, { key: "value" });
 
         await SessionFunctions.updateSessionDataInDatabase(s.helpers, res.session.handle, undefined);
 
-        let res4 = (await SessionFunctions.getSessionInformation(s.helpers, res.session.handle)).sessionData;
+        let res4 = (await SessionFunctions.getSessionInformation(s.helpers, res.session.handle)).sessionDataInDatabase;
         assert.deepStrictEqual(res4, {});
 
         await SessionFunctions.updateSessionDataInDatabase(s.helpers, res.session.handle, { key: "value 2" });
 
-        let res5 = (await SessionFunctions.getSessionInformation(s.helpers, res.session.handle)).sessionData;
+        let res5 = (await SessionFunctions.getSessionInformation(s.helpers, res.session.handle)).sessionDataInDatabase;
         assert.deepStrictEqual(res5, { key: "value 2" });
 
         await SessionFunctions.updateSessionDataInDatabase(s.helpers, res.session.handle, null);
 
-        let res6 = (await SessionFunctions.getSessionInformation(s.helpers, res.session.handle)).sessionData;
+        let res6 = (await SessionFunctions.getSessionInformation(s.helpers, res.session.handle)).sessionDataInDatabase;
         assert.deepStrictEqual(res6, {});
     });
 
@@ -782,27 +782,27 @@ describe(`session: ${printPath("[test/session.test.js]")}`, function () {
         let res = await SessionFunctions.createNewSession(s.helpers, "", false, {}, null);
 
         let res2 = await SessionFunctions.getSessionInformation(s.helpers, res.session.handle);
-        assert.deepStrictEqual(res2.sessionData, {});
+        assert.deepStrictEqual(res2.sessionDataInDatabase, {});
 
         await SessionFunctions.updateSessionDataInDatabase(s.helpers, res.session.handle, { key: "value" });
 
         let res3 = await SessionFunctions.getSessionInformation(s.helpers, res.session.handle);
-        assert.deepStrictEqual(res3.sessionData, { key: "value" });
+        assert.deepStrictEqual(res3.sessionDataInDatabase, { key: "value" });
 
         await SessionFunctions.updateSessionDataInDatabase(s.helpers, res.session.handle, undefined);
 
         let res4 = await SessionFunctions.getSessionInformation(s.helpers, res.session.handle);
-        assert.deepStrictEqual(res4.sessionData, {});
+        assert.deepStrictEqual(res4.sessionDataInDatabase, {});
 
         await SessionFunctions.updateSessionDataInDatabase(s.helpers, res.session.handle, { key: "value 2" });
 
         let res5 = await SessionFunctions.getSessionInformation(s.helpers, res.session.handle);
-        assert.deepStrictEqual(res5.sessionData, { key: "value 2" });
+        assert.deepStrictEqual(res5.sessionDataInDatabase, { key: "value 2" });
 
         await SessionFunctions.updateSessionDataInDatabase(s.helpers, res.session.handle, null);
 
         let res6 = await SessionFunctions.getSessionInformation(s.helpers, res.session.handle);
-        assert.deepStrictEqual(res6.sessionData, {});
+        assert.deepStrictEqual(res6.sessionDataInDatabase, {});
     });
 
     //check manipulating jwt payload
@@ -1133,7 +1133,7 @@ describe(`session: ${printPath("[test/session.test.js]")}`, function () {
         assert(typeof res2.status === "string");
         assert(res2.status === "OK");
         assert(typeof res2.userId === "string");
-        assert(typeof res2.sessionData === "object");
+        assert(typeof res2.sessionDataInDatabase === "object");
         assert(typeof res2.expiry === "number");
         assert(typeof res2.accessTokenPayload === "object");
         assert(typeof res2.timeCreated === "number");
@@ -1191,7 +1191,7 @@ describe(`session: ${printPath("[test/session.test.js]")}`, function () {
                             ...oI,
                             getSessionInformation: async (input) => {
                                 const info = await oI.getSessionInformation(input);
-                                info.sessionData = { test: 1 };
+                                info.sessionDataInDatabase = { test: 1 };
                                 return info;
                             },
                         }),