+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Renamed `getSessionData` to `getSessionDataFromDatabase` to clarify that it always hits the DB
 -   Renamed `updateSessionData` to `updateSessionDataInDatabase`
+-   Renamed `sessionData` to `sessionDataInDatabase` in `SessionInformation` and the input to `createNewSession`
 
 ## [13.1.3] - 2023-03-08